remove requirements from setup.py
even though, -e git+git://.. links can be given in requirements
files, this is not possible for the "install_requires" section
in setup.py. Therefore this is removed, and the requirements
have to be installed manually until either diggrtoolbox is
available via PyPI or I found out how to achieve that.